+++ plugins/decl_hdf5/CHANGELOG.md
@@ -19,6 +19,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 
 ### Fixed
+* Correctly delete files when using parallel HDF5 to prevent a crash
+  [428](https://gitlab.maisondelasimulation.fr/pdidev/pdi/-/issues/428)
 
 ### Security